# business-core/docs/api/business-core-openapi.yaml
openapi: 3.1.0
info:
  title: Business Core API
  version: "1.0.0"
  description: |
    API du Business Core — noyau métier générique intégré au kernel RT-Comops.

    Le Business Core expose un modèle de métier piloté par les métadonnées :
    le développeur déclare des Types Métier (le modèle), crée des Entreprises
    (les instances), et déclenche des Opérations. Le Business Core orchestre
    le kernel en façade.

    ## Authentification
    Chaque requête porte la clé Business Core du développeur via les en-têtes
    `X-BC-Client-Id` et `X-BC-Api-Key`. Le secret kernel n'est jamais exposé.

    ## Conventions
    - Versionnement de l'API dans l'URL (`/v1`), distinct de la version d'un Type Métier.
    - Opérations synchrones par défaut (`200`), différées en `202` avec une trace de suivi.
    - Erreurs au format RFC 7807 (Problem Details), enrichies de champs métier.
  contact:
    name: Équipe Business Core
  license:
    name: Usage académique
servers:
  - url: https://business-core.example.com
    description: Serveur du Business Core
security:
  - BusinessCoreKey: []
...
paths:
...
  /v1/business-types/{typeId}/versions/{versionNumber}/offers:
    parameters:
      - $ref: '#/components/parameters/TypeId'
      - $ref: '#/components/parameters/VersionNumber'
    post:
      tags: [Contenu de version]
      summary: Déclarer une définition d'offre
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOffer'
      responses:
        '201':
          description: Offre déclarée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Offer'
    get:
      tags: [Contenu de version]
      summary: Lister les offres types
      responses:
        '200':
          description: Liste des offres
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Offer'
  /v1/business-types/{typeId}/versions/{versionNumber}/offers/{offerId}:
    parameters:
      - $ref: '#/components/parameters/TypeId'
      - $ref: '#/components/parameters/VersionNumber'
      - name: offerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Contenu de version]
      summary: Consulter une offre
      responses:
        '200':
          description: L'offre
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Offer'
        '404':
          $ref: '#/components/responses/Problem'
    put:
      tags: [Contenu de version]
      summary: Modifier une offre
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOffer'
      responses:
        '200':
          description: Offre mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Offer'
        '404':
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Contenu de version]
      summary: Supprimer une offre
      description: Refusé (409) si un produit kernel est déjà mappé pour une entreprise.
      responses:
        '204':
          description: Offre supprimée
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
...
  /v1/business-types/{typeId}/versions/{versionNumber}/operations:
    parameters:
      - $ref: '#/components/parameters/TypeId'
      - $ref: '#/components/parameters/VersionNumber'
    post:
      tags: [Contenu de version]
      summary: Déclarer une opération et ses étapes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOperation'
      responses:
        '201':
          description: Opération déclarée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
...
  /v1/businesses:
    post:
      tags: [Entreprises]
      summary: Créer une entreprise
      description: Instancie un Type Métier à une version donnée. L'entreprise reste épinglée à cette version.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBusiness'
      responses:
        '201':
          description: Entreprise créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Business'
        '422':
          $ref: '#/components/responses/Problem'
    get:
      tags: [Entreprises]
      summary: Lister les entreprises
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Liste paginée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageBusiness'
...
  /v1/businesses/{businessId}/operations:
    parameters:
      - $ref: '#/components/parameters/BusinessId'
    get:
      tags: [Opérations]
      summary: Lister les opérations disponibles
      responses:
        '200':
          description: Opérations de l'entreprise
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Operation'
  /v1/businesses/{businessId}/operations/{name}:
    parameters:
      - $ref: '#/components/parameters/BusinessId'
      - name: name
        in: path
        required: true
        description: Nom de l'opération (ex. vente, reservation)
        schema:
          type: string
    get:
      tags: [Opérations]
      summary: Consulter une opération
      responses:
        '200':
          description: Détail de l'opération
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
        '404':
          $ref: '#/components/responses/Problem'
  /v1/businesses/{businessId}/operations/{name}:execute:
    parameters:
      - $ref: '#/components/parameters/BusinessId'
      - name: name
        in: path
        required: true
        description: Nom de l'opération à exécuter (ex. vente, reservation)
        schema:
          type: string
    post:
      tags: [Opérations]
      summary: Exécuter une opération
      description: |
        Déclenche un acte métier. Synchrone par défaut (`200` avec le résultat).
        Pour une opération différée, renvoie `202` avec une trace de suivi.
        Une règle bloquante renvoie `422` au format RFC 7807 enrichi.
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Clé d'idempotence pour éviter les doublons
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteOperation'
      responses:
        '200':
          description: Opération terminée (immédiate)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResult'
        '202':
          description: Opération acceptée, en cours (différée)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationPending'
        '422':
          description: Règle métier non respectée
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    BusinessCoreKey:
      type: apiKey
      in: header
      name: X-BC-Api-Key
      description: Clé d'API du développeur. Accompagnée de l'en-tête X-BC-Client-Id.
  parameters:
    TypeId:
      name: typeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VersionNumber:
      name: versionNumber
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    BusinessId:
      name: businessId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 20
        maximum: 100
  responses:
    Problem:
      description: Erreur au format RFC 7807
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      description: Détail d'erreur RFC 7807, enrichi de champs métier.
      properties:
        type:
          type: string
          example: /errors/regle-violee
        title:
          type: string
          example: Règle métier non respectée
        status:
          type: integer
          example: 422
        detail:
          type: string
          example: Un document ordonnance est requis pour cette vente.
        instance:
          type: string
          example: /v1/businesses/abc/operations/vente:execute
        violatedRule:
          type: string
          description: Code de la règle violée (extension métier)
        requiredAction:
          type: string
          enum: [BLOQUER, EXIGER, VALIDER, AJUSTER, ALERTER, DEROGER]
        requiredDocument:
          type: string
...
    CreateOffer:
      type: object
      required: [nom]
      properties:
        nom:
          type: string
          example: Médicament
        formePrix:
          type: string
          enum: [FIXE, GRATUIT, SUR_DEVIS]
        prix:
          type: number
        capacites:
          type: array
          items:
            type: string
            enum: [STOCKABLE, PLANIFIABLE, RESERVABLE, RECURRENT]
    Offer:
      allOf:
        - $ref: '#/components/schemas/CreateOffer'
        - type: object
          properties:
            id:
              type: string
              format: uuid
...
    CreateOperation:
      type: object
      required: [nom, etapes]
      properties:
        nom:
          type: string
          example: vente
        roleDeclencheur:
          type: string
        etapes:
          type: array
          items:
            type: object
            properties:
              ordre:
                type: integer
              typeEtape:
                type: string
                example: ENREGISTRER_VENTE
    Operation:
      allOf:
        - $ref: '#/components/schemas/CreateOperation'
        - type: object
          properties:
            id:
              type: string
              format: uuid
...
    CreateBusiness:
      type: object
      required: [typeId, versionNumber, nom]
      properties:
        typeId:
          type: string
          format: uuid
        versionNumber:
          type: integer
        nom:
          type: string
          example: Pharmacie du Centre
        organizationId:
          type: string
          format: uuid
          description: Référence Organization kernel (créée si absente)
...
    Business:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nom:
          type: string
        typeId:
          type: string
          format: uuid
        versionNumber:
          type: integer
        organizationId:
          type: string
          format: uuid
        cycleVie:
          type: string
          enum: [ACTIVE, SUSPENDUE, FERMEE]
...
    ExecuteOperation:
      type: object
      description: Entrées de l'opération (payload métier).
      properties:
        parametres:
          type: object
          additionalProperties: true
          description: Paramètres propres à l'opération (offre, quantité, bénéficiaire...)
      example:
        parametres:
          offreId: "a1b2"
          quantite: 2
          beneficiaireId: "c3d4"
    OperationResult:
      type: object
      properties:
        statut:
          type: string
          enum: [COMPLETEE]
        transactionId:
          type: string
          description: Référence de la transaction kernel produite
        traceId:
          type: string
          format: uuid
        details:
          type: object
          additionalProperties: true
    OperationPending:
      type: object
      properties:
        statut:
          type: string
          enum: [EN_COURS]
        traceId:
          type: string
          format: uuid
        suivi:
          type: string
          description: URL de suivi de la trace
          example: /v1/businesses/abc/traces/trace-4410
...
    PageInfo:
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
...
    PageBusiness:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Business'
        pageInfo:
          $ref: '#/components/schemas/PageInfo'
...